+++ src/seppl/__init__.py
@@ -1,4 +1,6 @@
-from ._plugin import Plugin, PluginWithLogging, OutputProducer, InputConsumer, check_compatibility, LoggingHandler, Initializable, init_initializable
+from ._plugin import Plugin, PluginWithLogging, OutputProducer, InputConsumer, check_compatibility
+from ._plugin import LoggingHandler, Initializable, init_initializable
+from ._plugin import AliasSupporter, get_aliases, has_aliases, get_all_names
 from ._args import split_cmdline, split_args, args_to_objects, is_help_requested, enumerate_plugins, escape_args, unescape_args, resolve_handler
 from ._registry import Registry, MODE_DYNAMIC, MODE_EXPLICIT, MODES
 from ._class_registry import ClassListerRegistry, get_class_lister, DEFAULT